scripts/feeds: return 1 to the shell if any of the feeds fail to update
Currently the feeds script always exit with the status 0, even if
it fails to update a feed. This patch changes this behaviour,
returning 1 to the shell if any of the feeds fail to update.

This keeps the current behaviour, i.e., try to update all feeds
even when one of them fails.

Signed-off-by: Jonh Wendell <jonh.wendell@oiwifi.com.br>